requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
gspread-asyncio==1.9.0
idna==3.4
magic-filter==1.0.11
multidict==6.0.4
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing_extensions==4.7.1
urllib3==1.26.16
yarl==1.9.2
